Add Airplanes to airplanes.rs
